Move the Source interface into signal, and rename other files to reflect
